+++ src/mandelbrot-helper.hpp
@@ -7,6 +7,6 @@
     Functions to help with the actual mandelbrot calcuations
 */
 
-double scale(int value, int fromLow, int fromHigh, int toLow, int toHigh);
+double scale(double value, double fromLow, double fromHigh, double toLow, double toHigh);
 int get_velocity(std::complex<double> c, int iterations);
 std::complex<double> mandelbrot(std::complex<double> z, std::complex<double> c);
